requirements.txt
psycopg2
python-dotenv
requests
apscheduler